requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
python-dotenv
werkzeug
# Flask Core
Flask==3.0.0
flask-cors==4.0.0

# Database
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-migrate==4.0.5

# Authentication
flask-jwt-extended==4.6.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
pandas==2.1.4
numpy==1.26.2

# Excel Support
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
flask-migrate
python-dotenv
psycopg2-binary
pandas
numpy
openpyxl
joblib
scikit-learn
